requirements.txt
streamlit>=1.25
bcrypt
python-dotenv
google-generativeai
